requirements.txt
flask
requests
pycryptodome
